Move silence utilities into their own file
