+++ src/components/alerting/AlertsDetailPage.tsx
@@ -49,9 +49,7 @@ import {
   SeverityBadge,
   SeverityHelp,
   silenceAlertURL,
-  SilenceTableRow,
   SourceHelp,
-  tableSilenceClasses,
 } from './AlertUtils';
 import { withFallback } from '../console/console-shared/error/error-boundary';
 import { SectionHeading } from '../console/utils/headings';
@@ -67,6 +65,7 @@ import {
   StatefulSetModel,
 } from '../console/models';
 import { Labels } from '../labels';
+import { SilenceTableRow, tableSilenceClasses } from './SilencesUtils';
 
 const AlertsDetailsPage_: React.FC<AlertsDetailsPageProps> = ({ history, match }) => {
   const { t } = useTranslation('plugin__monitoring-plugin');
